requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
audioread==3.0.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dominate==2.9.1
einops==0.7.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
grpcio==1.62.0
h5py==3.10.0
huggingface-hub==0.21.3
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
PySoundFile==0.9.0.post1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.3.0
timm==0.9.16
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
visdom==0.2.4
websocket-client==1.7.0
Werkzeug==3.0.1
zipp==3.17.0
